Move all files to _refactor folder before reestruture
