--- src/OFX.SDK/Attributes/OFXTagAttribute.cs
@@ -1,84 +1,0 @@
-﻿namespace OFX.SDK.Attributes;
-
-#region BSD-3 Copyright Information
-/*
-  Copyright (c) 2022-2024, Eric Roberto Darruiz
-  All rights reserved.
-  
-  Redistribution and use in source and binary forms, with or without
-  modification, are permitted provided that the following conditions are met:
-  
-  1. Redistributions of source code must retain the above copyright notice, this
-     list of conditions and the following disclaimer.
-  
-  2. Redistributions in binary form must reproduce the above copyright notice,
-     this list of conditions and the following disclaimer in the documentation
-     and/or other materials provided with the distribution.
-  
-  3. Neither the name of the copyright holder nor the names of its
-     contributors may be used to endorse or promote products derived from
-     this software without specific prior written permission.
-  
-  THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS"
-  AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
-  IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE
-  DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT HOLDER OR CONTRIBUTORS BE LIABLE
-  FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL
-  DAMAGES (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR
-  SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER
-  CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY,
-  OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
-  OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
-*/
-#endregion
-
-/// <summary>
-/// Represents an OFX tag for any OFX structure declarations.
-/// </summary>
-[AttributeUsage(AttributeTargets.Class | AttributeTargets.Property | AttributeTargets.Struct | AttributeTargets.Enum | AttributeTargets.Field,
-    Inherited = false, AllowMultiple = true)]
-public sealed class OFXTagAttribute : Attribute {
-    #region Ctor
-    /// <summary>
-    /// Initializes a new instace of the <see cref="OFXTagAttribute"/> class.
-    /// </summary>
-    /// <param name="specification">The <see cref="OFXSpecification"/> representing the specification version for the tag.</param>
-    /// <param name="name">The tag name.</param>
-    /// <param name="isHeaderTag">Define if the tag is a header tag. Defaults to <c>false</c>.</param>
-    public OFXTagAttribute(OFXSpecification specification, string name, bool isHeaderTag = false) {
-        if (string.IsNullOrWhiteSpace(name)) {
-            throw new ArgumentNullException(nameof(name));
-        }
-
-        Specification = specification;
-        Name = name;
-        IsHeaderTag = isHeaderTag;
-
-        if (!IsHeaderTag) {
-            if (!Name.StartsWith('<')) {
-                Name = $"<{Name}";
-            }
-            if (!Name.EndsWith('>')) {
-                Name = $"{Name}>";
-            }
-        }
-    }
-    #endregion
-
-    #region Properties
-    /// <summary>
-    /// Gets the specification version for the tag.
-    /// </summary>
-    public OFXSpecification Specification { get; private set; }
-
-    /// <summary>
-    /// Gets the tag name.
-    /// </summary>
-    public string Name { get; private set; }
-
-    /// <summary>
-    /// Gets or sets if the tag is a header tag.
-    /// </summary>
-    public bool IsHeaderTag { get; private set; }
-    #endregion
-}
