--- src/OFX.SDK/Specifications/Version102/Header/OFXHeaderSecurity.cs
@@ -1,52 +1,0 @@
-﻿namespace OFX.SDK.Specifications.Version102;
-
-#region BSD-3 Copyright Information
-/*
-  Copyright (c) 2022-2024, Eric Roberto Darruiz
-  All rights reserved.
-  
-  Redistribution and use in source and binary forms, with or without
-  modification, are permitted provided that the following conditions are met:
-  
-  1. Redistributions of source code must retain the above copyright notice, this
-     list of conditions and the following disclaimer.
-  
-  2. Redistributions in binary form must reproduce the above copyright notice,
-     this list of conditions and the following disclaimer in the documentation
-     and/or other materials provided with the distribution.
-  
-  3. Neither the name of the copyright holder nor the names of its
-     contributors may be used to endorse or promote products derived from
-     this software without specific prior written permission.
-  
-  THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS"
-  AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
-  IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE
-  DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT HOLDER OR CONTRIBUTORS BE LIABLE
-  FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL
-  DAMAGES (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR
-  SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER
-  CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY,
-  OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
-  OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
-*/
-#endregion
-
-/// <summary>
-/// Represents the application-level security, if any, that is used for the <c>&lt;OFX&gt;</c> block.
-/// </summary>
-/// <remarks>For more information about security, refer to Chapter 4, "Security." of documentation for the version 1.0.2.</remarks>
-[OFXVersion(OFXSpecification.Version102)]
-public enum OFXHeaderSecurity {
-    /// <summary>
-    /// Do not use any application-level security.
-    /// </summary>
-    [OFXValue(Constants.OFX_SECURITY_NONE)]
-    None = 0,
-
-    /// <summary>
-    /// Use Type 1 application-level security.
-    /// </summary>
-    [OFXValue(Constants.OFX_SECURITY_TYPE1)]
-    Type1 = 1
-}
