+++ src/OFX.SDK/_refactor/Specifications/Version102/ResponseMessageAggregates/SignonResponseMessage.cs
@@ -1,0 +1,48 @@
+using OFX.SDK._refactor.Specifications;
+using OFX.SDK._refactor.Specifications.Version102.Interfaces.ResponseMessageAggregates;
+
+namespace OFX.SDK._refactor.Specifications.Version102.ResponseMessageAggregates;
+
+#region BSD-3 Copyright Information
+/*
+  Copyright (c) 2022-2024, Eric Roberto Darruiz
+  All rights reserved.
+  
+  Redistribution and use in source and binary forms, with or without
+  modification, are permitted provided that the following conditions are met:
+  
+  1. Redistributions of source code must retain the above copyright notice, this
+     list of conditions and the following disclaimer.
+  
+  2. Redistributions in binary form must reproduce the above copyright notice,
+     this list of conditions and the following disclaimer in the documentation
+     and/or other materials provided with the distribution.
+  
+  3. Neither the name of the copyright holder nor the names of its
+     contributors may be used to endorse or promote products derived from
+     this software without specific prior written permission.
+  
+  THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS"
+  AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
+  IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE
+  DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT HOLDER OR CONTRIBUTORS BE LIABLE
+  FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL
+  DAMAGES (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR
+  SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER
+  CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY,
+  OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
+  OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
+*/
+#endregion
+
+/// <summary>
+/// Represents the Signon Response message aggregate. 
+/// Unlike other requests, the signon request <c>&lt;SONRS&gt;</c> does not appear within a transaction wrapper.
+/// </summary>
+/// <remarks>A client should use DTPROFUP and DTACCTUP only when the service provider that originated SONRS is the same 
+/// provider that is specified by SPNAME in the profile message set. A client can determine if the service provider is the 
+/// same by comparing the value of SPNAME in the appropriate message set with the value for SPNAME in the profile message set.</remarks>
+[OFXVersion(OFXSpecification.Version102)]
+[OFXTag(OFXSpecification.Version102, "SONRS")]
+public sealed record class SignonResponseMessage : IMessageResponse {
+}
